Top Panel Layout + Clear box button edited
Toppanel layout has been made so that the tf1 is centralised and the
clear box is on the left and enter(push) is on the right.
CLEAR has been changed to clear box to act as a way to clear tf1 as
previously there was no way to do so. Clear button on the operation button
area in the middlepanel is used to clear your entire history of inputted
numbers.

Signed-off-by: Duc Nguyen <ducn21@marvin.education.tas.gov.au>